Refactor and separte CGO and GO parts
Use Go build tags to separte CGO and GO parts of the library. Thus when
someone needs to use only GO parts (for testing puposes, etc.) he can
disable CGO by setting the CGO_ENABLED environment variable to 0 when
running the go tool.

For example:

	CGO_ENABLED=0 go test

REF: eawsy/aws-lambda-go#9
