+++ service/lambda/runtime/context.go
@@ -14,7 +14,7 @@
 // limitations under the License.
 //
 
-package context
+package runtime
 
 // CognitoIdentity provides information about the Amazon Cognito identity
 // provider when Lambda function invoked through the AWS Mobile SDK.
@@ -71,9 +71,6 @@ type ClientContext struct {
 	Environment map[string]string `json:"env"`
 }
 
-// RemainingTimeInMillisFunc prototype
-type RemainingTimeInMillisFunc func() int64
-
 // Context provides information about Lambda execution environment.
 //
 // For example, you can use the Context to determine the
@@ -126,12 +123,12 @@ type Context struct {
 	// It can be nil.
 	ClientContext *ClientContext `json:"client_context,omitempty"`
 
-	// RemainingTimeInMillis returns remaining execution time till the function will
-	// be terminated, in milliseconds.
+	// RemainingTimeInMillis returns remaining execution time till the function
+	// will be terminated, in milliseconds.
 	//
-	// The maximum time limit at which Lambda will terminate the function execution
-	// is set at the time the Lambda function is created. Information about the
-	// remaining time of function execution can be used to specify function behavior
-	// when nearing the timeout.
-	RemainingTimeInMillis RemainingTimeInMillisFunc `json:"-"`
+	// The maximum time limit at which Lambda will terminate the function
+	// execution is set at the time the Lambda function is created. Information
+	// about the remaining time of function execution can be used to specify
+	// function behavior when nearing the timeout.
+	RemainingTimeInMillis func() int64
 }
